requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
ecdsa==0.19.0
idna==3.7
passlib==1.7.4
pyasn1==0.6.0
pymongo==4.8.0
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.2
